+++ example/lib/example_widgets/multi_select_example.dart
@@ -64,7 +64,11 @@ class _MultiSelectDropdownState extends State<MultiSelectDropdown> {
               const Text('What are your favorite names?'),
               MultiSelectDropdownSearchFormField<String>(
                 textFieldConfiguration: TextFieldConfiguration(
-                  decoration: const InputDecoration(labelText: 'Names'),
+                  decoration: const InputDecoration(
+                    border: OutlineInputBorder(),
+                    hintText: 'Type to search',
+                    hintStyle: TextStyle(color: Colors.grey),
+                  ),
                   controller: _dropdownSearchFieldController,
                 ),
                 paginatedSuggestionsCallback: (pattern) async {
@@ -98,10 +102,32 @@ class _MultiSelectDropdownState extends State<MultiSelectDropdown> {
                   setState(() {});
                 },
                 suggestionsBoxController: suggestionBoxController,
-                validator: (value) => _selectedNames.isEmpty
-                    ? 'Please select at least one name'
-                    : null,
                 displayAllSuggestionWhenTap: true,
+                dropdownBoxConfiguration: DropdownBoxConfiguration(
+                    scrollbarConfiguration:
+                        ScrollbarConfiguration(thickness: 5),
+                    enabled: true,
+                    decoration: InputDecoration(
+                      border: OutlineInputBorder(),
+                      hintText: 'Selected Items',
+                      hintStyle: TextStyle(color: Colors.grey.shade600),
+                      errorBorder: OutlineInputBorder(
+                        borderSide: BorderSide(color: Colors.red),
+                      ),
+                      focusedErrorBorder: OutlineInputBorder(
+                        borderSide: BorderSide(color: Colors.red),
+                      ),
+                    ),
+                    chipConfiguration: ChipConfiguration(
+                      backgroundColor: Colors.orange,
+                      labelStyle: const TextStyle(color: Colors.white),
+                      deleteIcon: const Icon(Icons.close, color: Colors.white),
+                      deleteIconColor: Colors.white,
+                      side: const BorderSide(color: Colors.orange),
+                      shape: RoundedRectangleBorder(
+                        borderRadius: BorderRadius.circular(10),
+                      ),
+                    )),
               ),
               const Spacer(),
               ElevatedButton(