+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 1.1.2 - 11/02/2025
+- Added more support for Multi Select dropdown.
+- Configurations for Multi select dropdown.
+- Resolved issues requested by users.
+
 ## 1.1.1 - 09/12/2024
 - Suggestion box Scrollbar customization.
 - Resolved issues requested by users.